# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ResumeAI.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ProfileMetadata": {
      "get": {
        "tags": [
          "ProfileMetadata"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ProfileMetadata"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileMetadata"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileMetadata"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileMetadata"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ProfileMetadata"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ProfileMetadata/file-info": {
      "patch": {
        "tags": [
          "ProfileMetadata"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResumeFileInfoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResumeFileInfoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResumeFileInfoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProfileMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "nullable": true
          },
          "blob_path": {
            "type": "string",
            "nullable": true
          },
          "file_name": {
            "type": "string",
            "nullable": true
          },
          "file_size": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "file_url": {
            "type": "string",
            "nullable": true
          },
          "upload_date": {
            "type": "string",
            "format": "date-time"
          },
          "job_title": {
            "type": "string",
            "nullable": true
          },
          "current_company": {
            "type": "string",
            "nullable": true
          },
          "years_of_experience": {
            "type": "string",
            "nullable": true
          },
          "professional_bio": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "nullable": true
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "last_updated": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateResumeFileInfoRequest": {
        "type": "object",
        "properties": {
          "blobPath": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "format": "int32"
          },
          "fileUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter 'Bearer {token}' in the field below",
        "scheme": "Bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}